Move test for details js into own file
